% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpmgee.R
\name{cpmgee}
\alias{cpmgee}
\title{CPMs for clustered continuous response data based on GEE methods for ordinal data}
\usage{
cpmgee(
  formula,
  subjects,
  data,
  times,
  categories,
  corr.mod = "independence",
  alpha = 0.5,
  fit.opt = rep(NA, 5)
)
}
\arguments{
\item{formula}{an R formula object}

\item{subjects}{a character string specifying the name of the subject variable}

\item{data}{a data frame including response data and covariates}

\item{times}{a vector of the maximum times within clusters}

\item{categories}{a numeric value indicating the number of distinct values in response data}

\item{corr.mod}{a character string specifying the working correlation structure
("independence", "exchangeable", and "ar1")}

\item{alpha}{an initial value for the association parameter}

\item{fit.opt}{a vector of options to control the behavior of the fitting algorithm}
}
\value{
A list containing the following components:

\item{max.id}{number of clusters}

\item{fitted.values}{a vector of the fitted values}

\item{linear.predictors}{a vector of linear predictors}

\item{coefficients}{a vector of interecept and regression parameters}

\item{robust.var}{the robust (sandwich) variance matrix}

\item{alpha}{an estimate of the association parameter}

\item{grad1}{the	first derivative of generalized variance}

\item{grad2}{the	second derivative of generalized variance}
}
\description{
This function build the CPMs for clustered/longitudinal continuous response data
based on GEE methods for ordinal data.
}
\details{
We propose two feasible and computationally efficient approaches to
fit CPMs for clustered continuous response variables with different working correlation structures
(independence, exchangeable and AR1 working correlation structures).

CPMs with independence working correlation can be efficiently fit to clustered continuous responses
with thousands of distinct values based on CPMs and sandwich estimator for variances.

To improve efficiency, CPMs with more complex working correlation structures (exchangeable and AR1)
can be fit with a one-step GEE estimator for repolr (repeated measures proportional odds logistic regression
proposed by Parsons). The number of distinct respones values can be further reduced by
equal-quantile binning or rounding.

Estimates of the mean, quantiles, and exceedance probabilities onditional on covariates (new data)
can be derived from the model fit.
}
\examples{
data(data)

# independence working correlation structure
mod_cpmgee_ind <- cpmgee(formula = y ~ x + t, data = data, categories = length(unique(data$y)),
subjects = 'id', times = 1:6, corr.mod = 'independence', alpha = 0.5)

# exchangeable working correlation structure
mod_cpmgee_ex <- cpmgee(formula = y ~ x + t, data = data, categories = length(unique(data$y)),
subjects = 'id', times = 1:6, corr.mod = 'exchangeable', alpha = 0.5)

# new data
new_data <- data.frame(x = c(0,1), t = 0.2)

# conditional quantities for independence working correlation structure
mean_ind <- mean_cpmgee(mod_cpmgee_ind, data$y, new_data)
median_ind <- quantile_cpmgee(mod_cpmgee_ind, data$y, new_data, 0.5)
cdf_ind <- cdf_cpmgee(mod_cpmgee_ind, data$y, new_data, 5)

# conditional quantities for exchangeable working correlation structure
mean_ex <- mean_cpmgee(mod_cpmgee_ex, data$y, new_data)
median_ex <- quantile_cpmgee(mod_cpmgee_ex, data$y, new_data, 0.5)
cdf_ex <- cdf_cpmgee(mod_cpmgee_ex, data$y, new_data, 0.5)
}
\references{
Tian et al. "Analyzing clustered continuous response variables with ordinal
 regression models." (2022) (to be submitted)

Parsons, N. (2017). repolr: an R package for fitting proportional-odds models to repeated
ordinal scores. R package version 3.4 https://CRAN.R-project.org/package=repolr

Harrell, F. (2020). rms: Regression modeling strategies. R package version 6.1.0. https://CRAN.R-project.org/package=rms
}
\seealso{
\code{\link{cdf_cpmgee}, \link{quantile_cpmgee}, \link{mean_cpmgee}}
}
