% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icormat.R
\name{icormat}
\alias{icormat}
\title{Obtain the working correlation matrix}
\usage{
icormat(mod, smat, modtype = "gee", diffmeth, alpha, corrmod, h)
}
\arguments{
\item{mod}{the fitted model with independence working correlation structure}

\item{smat}{the matrix of correlations among elements in $Z_{it}$}

\item{modtype}{a character string specifying the model type (CPM or GLM)}

\item{diffmeth}{a character string specifying the method used for estimation of alpha}

\item{alpha}{the association parameter}

\item{corrmod}{the specified working correlation structure}

\item{h}{a numeric value of options to control the fitting algorithm}
}
\value{
the working correlation matrix
}
\description{
Obtain the working correlation matrix
}
