% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_functions.R
\name{mean_orm}
\alias{mean_orm}
\title{Calculate expectations conditional on covariates (independence working correlation structure)}
\usage{
mean_orm(mod, cont_y, new.data, se = TRUE)
}
\arguments{
\item{mod}{the model}

\item{cont_y}{the continuous response data}

\item{new.data}{the new data}

\item{se}{if confidence intervals needed (default = TRUE)}
}
\value{
A list containing the following components:

\item{est}{a vector of estimated condtional expectations}

\item{se}{a vector of estimated standard errors}

\item{lb}{a vector of estimated lower bounds of 95\% confidence intervals}

\item{ub}{a vector of estimated upper bounds of 95\% confidence intervals}
}
\description{
This functions calculates the expectations conditional on covariates based on the fitted model
(with independence working correlation structure) and new data.
}
