% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upalpha.R
\name{upalpha}
\alias{upalpha}
\title{Update the association parameter}
\usage{
upalpha(hgmat, alpha, diffmeth, h)
}
\arguments{
\item{hgmat}{a list of H and G matrices (see reference for repolr for more details)}

\item{alpha}{a numeric value for current alpha value}

\item{diffmeth}{a character string specifying the method used for estimation of alpha}

\item{h}{a numeric value of finite differencing}
}
\value{
A list containing the following components:

\item{gvb}{the first derivative of generalized variance at convergence}

\item{ggvb}{the second derivative of generalized variance at convergence}

\item{alpha}{the updated assocation parameter}
}
\description{
This function updates alpha, the association parameter
}
