\name{cpmgee-package}
\alias{cpmgee-package}
\alias{cpmgee}
\docType{package}
\title{
  CPMs for Clustered Continuous Response Variables Based on GEE Methods
\description{
  
}
\details{
  
}
\author{
Yuqi Tian

Maintainer: Yuqi Tian <yuqi.tian@vanderbilt.edu>
}
\references{
  
}
\keyword{ package }
\seealso{
 
}
\examples{
  
}
