% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ord.expand.R
\name{ord.expand}
\alias{ord.expand}
\title{Expand data for model fitting}
\usage{
ord.expand(space, formula, times, poly, data, subjects, categories)
}
\arguments{
\item{space}{a vector indicating the category spacing when fitting the polynomial model}

\item{formula}{an R formula}

\item{times}{a vector of times which occur within subject clusters}

\item{poly}{a numeric variable indicating the order of the polynomial contrasts used for the cut-point model}

\item{data}{a data frame including response data and covariates}

\item{subjects}{a character string specifying the subject variable}

\item{categories}{a numeric variable indicating the number of ordinal levels}
}
\value{
an expanded data frame
}
\description{
This function expand each response value into indicator for model fitting
}
